fix: CSV 정수 필드 파싱 오류 수정 (8개 모델 적재 실패 해결)
## 문제
- XGBoost: pred_direction = "0" (정수 형식) → 성공 ✅
- LightGBM: pred_direction = "1.0" (실수 형식) → 실패 ❌
- Integer.parseInt("1.0") → NumberFormatException 발생

## 영향받은 모델 (8개)
- LightGBM, CatBoost, RandomForest, AdaBoost
- GradientBoosting, HistGradientBoosting
- LogisticRegression, StackingEnsemble

## 해결책
predDirection과 correct 필드를 actualDirection과 동일하게 파싱:
- 변경 전: Integer.parseInt(nextLine[5].trim())
- 변경 후: (int) Double.parseDouble(nextLine[5].trim())

## 결과
- 기존: 4개 모델만 성공 (3,684건)
- 수정 후: 12개 모델 모두 성공 (10,512건 예상)

## 기술적 배경
- CSV 파일 생성 시 Python에서 int/float 타입이 섞임
- XGBoost는 정수로 저장, 다른 모델은 실수로 저장
- Double.parseDouble()은 정수/실수 모두 처리 가능

+++ src/main/java/com/coinreaders/engine/application/AiPredictionDataService.java
@@ -187,12 +187,12 @@ public class AiPredictionDataService {
                     BigDecimal actualReturn = new BigDecimal(nextLine[2].trim());
                     BigDecimal takeProfitPrice = new BigDecimal(nextLine[3].trim());
                     BigDecimal stopLossPrice = new BigDecimal(nextLine[4].trim());
-                    Integer predDirection = Integer.parseInt(nextLine[5].trim());
+                    Integer predDirection = (int) Double.parseDouble(nextLine[5].trim()); // 1.0 → 1 (모델별 형식 통일)
                     BigDecimal predProbaUp = new BigDecimal(nextLine[6].trim());
                     BigDecimal predProbaDown = new BigDecimal(nextLine[7].trim());
                     BigDecimal maxProba = new BigDecimal(nextLine[8].trim());
                     BigDecimal confidence = new BigDecimal(nextLine[9].trim());
-                    Integer correct = Integer.parseInt(nextLine[10].trim());
+                    Integer correct = (int) Double.parseDouble(nextLine[10].trim()); // 1.0 → 1 (모델별 형식 통일)
 
                     HistoricalAiPrediction prediction = HistoricalAiPrediction.of(
                         MARKET,
